requirements.txt
streamlit 
pandas 
opencv-python 
Pillow 
numpy 
easyocr 
matplotlib